Apply rustfmt and fix Clippy warnings

+++ rand_distr/tests/pdf.rs
@@ -57,7 +57,7 @@ fn normal() {
 
     let mut diff = [0.; HIST_LEN];
     for (i, n) in hist.normalized_bins().enumerate() {
-        let bin = (n as f64) / (N_SAMPLES as f64);
+        let bin = n / (N_SAMPLES as f64);
         diff[i] = (bin - expected[i]).abs();
     }
 
@@ -140,7 +140,7 @@ fn skew_normal() {
 
     let mut diff = [0.; HIST_LEN];
     for (i, n) in hist.normalized_bins().enumerate() {
-        let bin = (n as f64) / (N_SAMPLES as f64);
+        let bin = n / (N_SAMPLES as f64);
         diff[i] = (bin - expected[i]).abs();
     }
